moved the preliminary Octree code to the ffx.potential.nonbonded.octree package; added code to build amino acid carbonyl oxygen atoms; updated the ffx.ui.ModelingShell to not delete systems after running Groovy scripts if the FFX GUI is running
