+++ modules/potential/src/main/java/ffx/potential/nonbonded/octree/OctreeParticle.java
@@ -1,4 +1,4 @@
-package ffx.potential.nonbonded.implicit;
+package ffx.potential.nonbonded.octree;
 
 /**
  * OctreeParticle:
