+++ modules/potential/src/main/java/ffx/potential/nonbonded/octree/OctreePoint.java
@@ -1,4 +1,4 @@
-package ffx.potential.nonbonded.implicit;
+package ffx.potential.nonbonded.octree;
 
 import java.util.logging.Logger;
